requirements.txt
litellm>=1.0.0
python-dotenv>=1.0.0
openai>=1.5.0
tiktoken>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
httpx>=0.25.0
lxml>=4.9.0
mem0ai
langgraph
rich>=13.6.0
arxiv
browsergym
pyan3
openhands-aci==0.3.1
docker==7.1.0
python-docx==1.1.2
pylatexenc==2.10

# PDF processing and vector database dependencies
langchain-community==0.3.8
langchain-text-splitters==0.3.5
faiss-cpu==1.11.0
langchain-openai==0.2.9
langchain-aws==0.2.1
huggingface-hub==0.28.1
langchain==0.3.9
langchain-anthropic==0.3.0
boto3==1.37.7
faiss-cpu==1.11.0
mlflow-skinny==2.14.1
tornado
